Moved ValuePart struct inside EvidenceConverter
