+++ tracer/src/Datadog.Trace/Iast/EvidenceConverter.cs
@@ -8,6 +8,7 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using System.Linq;
 using Datadog.Trace.Iast.Helpers;
 using Datadog.Trace.Vendors.Newtonsoft.Json;
 
@@ -80,7 +81,7 @@ internal class EvidenceConverter : JsonConverter<Evidence>
         writer.WriteEndObject();
     }
 
-    internal static void ToJsonTaintedValue(JsonWriter writer, string value, Range[] ranges)
+    private static void ToJsonTaintedValue(JsonWriter writer, string value, Range[] ranges)
     {
         writer.WriteStartArray();
         int start = 0;
@@ -104,7 +105,7 @@ internal class EvidenceConverter : JsonConverter<Evidence>
         writer.WriteEndArray();
     }
 
-    internal static void WriteValuePart(JsonWriter writer, string? value, Source? source = null)
+    private static void WriteValuePart(JsonWriter writer, string? value, Source? source = null)
     {
         if (value == null) { return; }
         writer.WriteStartObject();
@@ -119,7 +120,7 @@ internal class EvidenceConverter : JsonConverter<Evidence>
         writer.WriteEndObject();
     }
 
-    internal static void WriteRedactedValuePart(JsonWriter writer, string? value, Source? source = null)
+    private static void WriteRedactedValuePart(JsonWriter writer, string? value, Source? source = null)
     {
         writer.WriteStartObject();
         writer.WritePropertyName("redacted");
@@ -139,7 +140,7 @@ internal class EvidenceConverter : JsonConverter<Evidence>
         writer.WriteEndObject();
     }
 
-    internal static void Write(JsonWriter writer, ValuePart part)
+    private static void Write(JsonWriter writer, ValuePart part)
     {
         if (part.IsRedacted)
         {
@@ -175,6 +176,116 @@ internal class EvidenceConverter : JsonConverter<Evidence>
         }
 
         writer.WriteEndArray();
+    }
+
+    private readonly struct ValuePart
+    {
+        public ValuePart(string value) // String value part
+        {
+            this.Value = value;
+            this.Source = null;
+            this.SensitiveRanges = null;
+            this.IsRedacted = false;
+        }
+
+        public ValuePart(Source? source) // Redacted value part
+        {
+            this.Value = null;
+            this.Source = source;
+            this.SensitiveRanges = null;
+            this.IsRedacted = true;
+        }
+
+        public ValuePart(string value, Source? source, bool isRedacted = false) // Non redacted value part
+        {
+            this.Value = value;
+            this.Source = source;
+            this.SensitiveRanges = null;
+            this.IsRedacted = isRedacted;
+        }
+
+        public ValuePart(string value, Range range, LinkedList<Range> intersections) // Redactable value part
+        {
+            this.Value = value;
+            this.Source = range.Source;
+
+            // shift ranges to the start of the tainted range and sort them
+            this.SensitiveRanges = new LinkedList<Range>(intersections.Select(r => r.Shift(-range.Start)).OrderBy(r => r.Start));
+            this.IsRedacted = false;
+        }
+
+        public string? Value { get; }
+
+        public Source? Source { get; }
+
+        public bool IsRedacted { get; }
+
+        public LinkedList<Range>? SensitiveRanges { get; }
+
+        public bool ShouldRedact => !IsRedacted && (SensitiveRanges != null || (Source != null && Source.IsSensitive));
+
+        public List<ValuePart> Split()
+        {
+            var parts = new List<ValuePart>();
+            if (Value != null)
+            {
+                if (Source != null && Source.IsSensitive)
+                {
+                    // redact the full tainted value as the source is sensitive (password, certificate, ...)
+                    AddValuePart(0, Value.Length, true, parts);
+                }
+                else if (SensitiveRanges != null)
+                {
+                    // redact only sensitive parts
+                    int index = 0;
+                    foreach (var sensitive in SensitiveRanges)
+                    {
+                        var start = sensitive.Start;
+                        var end = sensitive.Start + sensitive.Length;
+                        // append previous tainted chunk (if any)
+                        AddValuePart(index, start, false, parts);
+                        // append current sensitive tainted chunk
+                        AddValuePart(start, end, true, parts);
+                        index = end;
+                    }
+
+                    // append last tainted chunk (if any)
+                    AddValuePart(index, Value.Length, false, parts);
+                }
+            }
+
+            return parts;
+        }
+
+        private void AddValuePart(int start, int end, bool redact, List<ValuePart> valueParts)
+        {
+            if (start < end)
+            {
+                var chunk = Value!.Substring(start, end - start);
+                if (!redact)
+                {
+                    // append the value
+                    valueParts.Add(new ValuePart(chunk, Source));
+                }
+                else
+                {
+                    var length = chunk.Length;
+                    var matching = Source!.Value!.IndexOf(chunk);
+                    if (matching >= 0)
+                    {
+                        // if matches append the matching part from the redacted value
+                        var pattern = Source!.RedactedValue!.Substring(matching, length);
+                        valueParts.Add(new ValuePart(pattern, Source, true));
+                    }
+                    else
+                    {
+                        // otherwise redact the string
+                        var pattern = Source.RedactString(chunk);
+                        valueParts.Add(new ValuePart(pattern, Source, true));
+                    }
+                }
+            }
+        }
     }
 
     private class ValuePartIterator : IEnumerable<ValuePart?>
